moved close to separate script
