+++ hatyan/__init__.py
@@ -24,27 +24,3 @@ from hatyan.kw_overschrijding import *
 
 import warnings
 warnings.filterwarnings(action='always', category=DeprecationWarning)
-
-
-import matplotlib.pyplot as plt
-def close(fig=None):
-    """
-    wrapper around matplotlib.pyplot.close() to enable closing of plots generated by hatyan without explicitly importing matplotlib
-
-    Parameters
-    ----------
-    fig : None or int or str or .Figure
-        The figure to close. There are a number of ways to specify this:
-        None: the current figure
-        .Figure: the given .Figure instance
-        int: a figure number
-        str: a figure name
-        'all': all figures.
-
-    Returns
-    -------
-    None.
-
-    """
-    
-    plt.close(fig)
